get_active_span_context {otel}	R Documentation

Returns the active span context

Description

This is sometimes useful for logs or metrics, to associate logging and metrics reporting with traces.

Usage

get_active_span_context()

Details

Note that logs and metrics instruments automatically use the current span context, so often you don't need to call this function explicitly.

Value

The active span context, an otel_span_context object. If there is no active span context, then an invalid span context is returned, i.e. spc$is_valid() will be FALSE for the returned spc.

Examples

fun <- function() {
  otel::start_local_active_span("fun")
  fun2()
}
fun2 <- function() {
  otel::log("Log message", span_context = otel::get_active_span_context())
}
fun()

[Package otel version 0.2.0 Index]