log_severity_levels {otel}	R Documentation

OpenTelemetry log severity levels

Description

A named integer vector, the severity levels in numeric form. The names are the severity levels in text form. otel functions accept both forms as severity levels, but the text form is more readable.

Value

Not applicable.

See Also

Other OpenTelemetry logs API: is_logging_enabled(), log()

Examples

log_severity_levels

[Package otel version 0.2.0 Index]