OpenTelemetry Up-Down Counter Object

otel_meter_provider -> otel_meter -> otel_counter, otel_up_down_counter, otel_histogram, otel_gauge

Value

Not applicable.

Details

Usually you do not need to deal with otel_up_down_counter objects directly. up_down_counter_add() automatically sets up a meter and creates an up-down counter instrument, as needed.

An up-down counter object is created by calling the create_up_down_counter() method of an otel_meter_provider().

You can use the add() method to increment or decrement the counter.

In R up-down counters are represented by double values.

Methods

up_down_counter$add()

Increment or decrement the up-down counter by a fixed amount.

Usage

up_down_counter$add(value, attributes = NULL, span_context = NULL, ...)

Arguments

value: Value to increment of decrement the up-down counter with.

attributes: Additional attributes to add.

span_context: Span context. If missing, the active context is used, if any.

Value

The up-down counter object itself, invisibly.

See also

Examples

mp <- get_default_meter_provider()
mtr <- mp$get_meter()
ctr <- mtr$create_up_down_counter("session")
ctr$add(1)