tracing-constants {otel}	R Documentation

OpenTelemetry tracing constants

Description

Various constants related OpenTelemetry tracing.

Usage

invalid_trace_id

invalid_span_id

span_kinds

span_status_codes

Details

invalid_trace_id

invalid_trace_id is a string scalar, an invalid trace id. If there is no active span, then get_active_span_context() returns a span context that has an invalid trace id.

invalid_span_id

invalid_span_id is a string scalar, an invalid span id. If there is no active span, then get_active_span_context() returns a span context that has an invalid span id.

span_kinds

span_kinds is a character vector listing all possible span kinds. See the OpenTelemetry specification for when to use which.

span_status_codes

span_status_codes is a character vector listing all possible span status codes. You can set the status code of a a span with the set_status() method of otel_span objects. If not set explicitly, and the span is ended automatically (by start_local_active_span(), local_active_span() or with_active_span()), then otel sets the status automatically to "ok" or "error", depending on whether the span ended during handling an error.

Value

Not applicable.

See Also

Other OpenTelemetry trace API: Zero Code Instrumentation, end_span(), is_tracing_enabled(), local_active_span(), start_local_active_span(), start_span(), with_active_span()

Examples

invalid_trace_id
invalid_span_id
span_kinds
span_status_codes

[Package otel version 0.2.0 Index]