Start a new OpenTelemetry span, using the default tracer

A span represents a unit of work or operation. Spans are the building blocks of traces. start_span() is typically called at the beginning of a function or the beginning of some larger operation, e.g. an HTTP request.

Usage

start_span(
  name = NULL,
  tracer_name = NULL,
  attributes = NULL,
  links = NULL,
  options = NULL,
  scope = parent.frame(),
  activation_scope = parent.frame()
)

Arguments

name

Name of the span. If not specified it will be "<NA>".

tracer_name

The name of the tracer to use, see get_tracer().

attributes

Span attributes. You may use as_attributes() to convert R objects to OpenTelemetry attributes. OpenTelemetry supports the following R types as attributes: character, logical, double, integer.

links

TODO

options

TODO

scope

TODO

activation_scope

TODO

Value

The new OpenTelemetry span object (of class otel_span), invisibly. See otel_span for information about the returned object.

Details

Calls get_tracer() with tracer_name as the tracer name, and then calls the $start_span() method of the tracer with the other arguments.

Lifetime of a span

TODO

The active span

TODO

Span attributes

TODO

Span links

TODO

Span options

TODO

See also